+++ core/src/com/pi4/mgmtgame/resources/Grain.java
@@ -6,7 +6,7 @@ public abstract class Grain extends Resources implements Serializable {
 	static int basePrice = 1;
 	int growingTime = 1;
 
-	public abstract int getGrowingTime();
+	public abstract float getGrowingTime();
 	public abstract Plant getGrownPlant();
 	public abstract int getId();
 	public abstract String getFieldSpriteName();
